implement a proper solution for generating json schemas
