+++ packages/config/scripts/generate-json-schema.ts
@@ -1,0 +1,3 @@
+import { generateJsonSchema } from "../src/json-schema.ts";
+
+await generateJsonSchema();
